[NCL-6167] Use new PME/DA endpoint parameters
The new PME (4.3) and DA (2.x) introduces 2 new parameters:

- brewPullActive (default: false)
- mode (default: "")

The 'brewPullActive' flag is set on whether to also consider artifacts
from Brew during alignment.

The mode is used to specify to which aligned versions to align to. Currently
there are 4 modes:

- PERSISTENT (permanent aligned builds)
- TEMPORARY (temporary aligned builds)
- SERVICE (service aligned builds)
- SERVICE-TEMPORARY (temporary service aligned builds)

Those modes and the corresponding aligned version suffixes are
configured on the Dependency Analysis service.